Tweaked dap config
Removed DapMason in favour of install debuggers locally. Tweaked the
debug config for python to directly add an adapter and configs for
testing and execution rather than rely on the python dap plugin.

So many tutorials out there for nvim set up which are incomplete
or give bad results.

Still need to update the readme. It's on my todo list!